% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detrending.R
\name{linear.detrend}
\alias{linear.detrend}
\title{A function to estimate linear trends}
\usage{
linear.detrend(data)
}
\arguments{
\item{data}{A numeric vector (e.g. a time series), from which the trend component is to be estimated}
}
\value{
A character vector of length(data) that contains the trend components.
}
\description{
Estimates a linear trend.
}
\details{
This function calculates a linear trend component to use for spatio-temporal normalization.
}

