% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize.spatiotemporal.cube.R
\name{normalize.spatiotemporal.cube}
\alias{normalize.spatiotemporal.cube}
\title{A function to normalize a spatio-temporal data cube}
\usage{
normalize.spatiotemporal.cube(data, data.trend=NA, SUBTRENDSD=F, TRENDCOR=F, ref.idx = c(1,30))
}
\arguments{
\item{data}{3-dimensional array of the form longiutde-latitude-time}

\item{data.trend}{(optional): 3-dimensional array with trend components in the form longitude-latitude-time or NA}

\item{SUBTRENDSD}{Logical, should the trend be subtracted before computing the standard deviation in the reference period?}

\item{TRENDCOR}{Logical, should the normalization correct for trends in the out-of-base period?}

\item{ref.idx}{Numeric Vector of indices that specify the reference period (in time)}
}
\value{
List with three elements:
$data.original: original data array (3D-array)
$data.norm: normalized array using the conventionally applied normalization
$data.norm.cor: normalized array using the proposed correction
}
\description{
Normalization of a spatio-temporal data cube based on a reference period and correction for induced artefacts.
}
\details{
This function returns a normalized spatiotemporal data cube that can be compared across space and time, given that prerequisites are fulfilled (Gaussian data, stationarity in the reference period).
}
\author{
Sebastian Sippel
}
\references{
Sippel et al, (2015) An accurate quantification of climate variability and extremes. Submitted.
}

