% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detrending.R
\name{SSA.detrend}
\alias{SSA.detrend}
\title{A function to estimate nonlinear trends}
\usage{
SSA.detrend(data, borders.wl, M, n.comp)
}
\arguments{
\item{data}{A numeric vector (e.g. a time series), from which the trend component is to be estimated}

\item{borders.wl}{See ?filterTSeriesSSA}

\item{M}{See ?filterTSeriesSSA}

\item{n.comp}{See ?filterTSeriesSSA}
}
\value{
A character vector of length(data) that contains the trend components.
}
\description{
Estimate the nonlinear trend components of a time series using Singular Spectrum Analysis. A wrapper function around filterTSeriesSSA (spectral.methods)
}
\details{
This function calculates the nonlinear trend component. Currently, no padding is implemented, but will come soon.
}

